backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
google.genai
googletrans==4.0.0-rc1
firebase-admin
python-dotenv